requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
HyperPyYAML==1.2.2
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.24
langchain-text-splitters==0.3.8
langsmith==0.4.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
openai==1.90.0
orjson==3.10.18
packaging==24.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scipy==1.15.3
sentencepiece==0.2.0
sniffio==1.3.1
soundfile==0.13.1
speechbrain==1.0.3
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.1
torchaudio==2.7.1
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
youtube-transcript-api==1.1.0
zstandard==0.23.0
